Adequação site com pdvprodu

+++ application/views/home.php
@@ -78,19 +78,22 @@ else {
 		} else {
 			echo ('<div class="col-xs-6 col-sm-3">
 					<div class="thumbnail">');
-						list($width, $height, $type, $attr) = getimagesize(base_url("includes/images/produtos/thumbnail/".$reg->codigo.".png"));
-						if ($height > $width)
-							echo ('<img src="'. base_url("includes/images/produtos/thumbnail/".$reg->codigo.".png") .'" style="height: 170px !important;" alt="'.$reg->nome.'">');
-						else
-							echo ('<img src="'. base_url("includes/images/produtos/thumbnail/".$reg->codigo.".png") .'" style="width: 170px !important;" alt="'.$reg->nome.'">');
+						if (file_exists(base_url("includes/images/produtos/thumbnail/".$reg->codigo.".png"))) {
+							list($width, $height, $type, $attr) = getimagesize(base_url("includes/images/produtos/thumbnail/".$reg->codigo.".png"));
+							if ($height > $width)
+								echo ('<img src="'. base_url("includes/images/produtos/thumbnail/".$reg->codigo.".png") .'" style="height: 170px !important;" alt="'.$reg->nome.'">');
+							else
+								echo ('<img src="'. base_url("includes/images/produtos/thumbnail/".$reg->codigo.".png") .'" style="width: 170px !important;" alt="'.$reg->nome.'">');
+						} else
+							echo ('<img src="'. base_url("includes/images/produtos/semimagem.png") .'" style="width: 170px !important;" alt="'.$reg->nome.'">');
 						
 						echo ('<div class="codigo">Codigo: '.$reg->codigo.'</div>
 						<div class="caption">
 							<h3>'.formataString($reg->nome).'</h3>');
-							if (($reg->prepro < $reg->pvenda) && ($reg->prepro > 0))
-								echo ('<p><strike>De: R$ '.number_format($reg->prepro,2,","," ").'</strike> Por:  R$ '.number_format($reg->pvenda,2,","," ").'</p>');
+							if (($reg->prepro < $reg->preco) && ($reg->prepro > 0))
+								echo ('<p><strike>De: R$ '.number_format($reg->preco,2,","," ").'</strike> Por:  R$ '.number_format($reg->prepro,2,","," ").'</p>');
 							else
-								echo ('<p>Por:  R$ '.number_format($reg->pvenda,2,","," ").'</p>');
+								echo ('<p>Por:  R$ '.number_format($reg->preco,2,","," ").'</p>');
 							echo ('<p><a href="'. base_url("Produtos/detalhes/".formataStringToURL($reg->nome)."/".$reg->codigo) .'" class="btn btn-default" role="button"><span class="glyphicon glyphicon-plus"> </span> Detalhes</a> <a href="#" class="btn btn-primary" role="button"><span class="glyphicon glyphicon-shopping-cart"> </span> Comprar</a> </p>
 						</div>
 					</div>
